rm whitespace

+++ SPM-Grupparbete Unity/Assets/Scripts/DamageEvent_DieEvents/DamageDealt.cs
@@ -7,10 +7,10 @@ namespace EgilScripts.DieEvents
     public class DamageDealt : Event
     {
         // hur mycket skada ska jag göra på det gameobjektet jag träffade.
-        public int amountOfDamage;
+        public float amountOfDamage;
         public GameObject gameObject;
  
-        public DamageDealt(GameObject gameObject, int amountOfDamage)
+        public DamageDealt(GameObject gameObject, float amountOfDamage)
         {
             this.amountOfDamage = amountOfDamage;
             this.gameObject = gameObject;
